+++ packages/stacked/example/lib/ui/home/home_viewmodel.dart
@@ -1,13 +1,22 @@
+import 'package:new_architecture/app/app.locator.dart';
 import 'package:new_architecture/app/app.logger.dart';
+import 'package:new_architecture/app/app.router.dart';
 import 'package:stacked/stacked.dart';
+import 'package:stacked_services/stacked_services.dart';
 
 class HomeViewModel extends BaseViewModel {
   final log = getLogger('HomeViewModel');
+  final NavigationService _navigationService = locator<NavigationService>();
+
   HomeViewModel() {
     log.d('created');
   }
 
   String title = 'default';
+
+  void navigate() {
+    _navigationService.navigateTo(Routes.nonReactiveView);
+  }
 
   void initialise() {
     log.i('initialise');