+++ packages/stacked_generator/CHANGELOG.md
@@ -1,7 +1,7 @@
 # ChangeLog
 
 ## 0.4.2
-
+- Adds `initialValue` parameter for `FormTextField` to support initial value in the `TextEditingController`
 - Fixes `Route TransitionsBuilders` generation
 
 ## 0.4.1