+++ packages/stacked_generator/CHANGELOG.md
@@ -1,8 +1,13 @@
 # ChangeLog
 
-## 0.4.2
+## 0.4.3
 
 - Added `Environment` on Dependency Injection
+
+## 0.4.2
+
+- Adds `initialValue` parameter for `FormTextField` to support initial value in the `TextEditingController`
+- Fixes `Route TransitionsBuilders` generation
 
 ## 0.4.1